refactor: extract application interface to services
Extracting the application interface to packages/services so that
it can be extended and referenced by other packages.

Closes https://app.asana.com/0/1202173912993191/1202412897951248/f

userclientinterface

deinitmode

application interface
