+++ packages/services/src/Domain/Application/DeinitCallback.ts
@@ -1,0 +1,5 @@
+import { DeinitSource } from './DeinitSource'
+import { DeinitMode } from './DeinitMode'
+import { AppGroupManagedApplication } from './ApplicationInterface'
+
+export type DeinitCallback = (application: AppGroupManagedApplication, mode: DeinitMode, source: DeinitSource) => void
