+++ packages/services/src/Domain/index.ts
@@ -1,6 +1,11 @@
 export * from './Alert/AlertService'
 export * from './Api/ApiServiceInterface'
 export * from './Application/ApplicationStage'
+export * from './Application/DeinitCallback'
+export * from './Application/DeinitSource'
+export * from './Application/DeinitMode'
+export * from './Application/UserClientInterface'
+export * from './Application/ApplicationInterface'
 export * from './Challenge'
 export * from './Device/DesktopDeviceInterface'
 export * from './Device/DesktopWebCommunication'
