+++ packages/snjs/lib/ApplicationGroup/AppGroupCallback.ts
@@ -1,5 +1,4 @@
-import { AppGroupManagedApplication } from './../Application/ApplicationInterface'
-import { DeviceInterface } from '@standardnotes/services'
+import { AppGroupManagedApplication, DeviceInterface } from '@standardnotes/services'
 import { ApplicationDescriptor } from './ApplicationDescriptor'
 
 export type AppGroupCallback<D extends DeviceInterface = DeviceInterface> = {
