+++ packages/snjs/lib/ApplicationGroup/DeinitCallback.ts
@@ -1,5 +1,3 @@
-import { DeinitSource } from '../Types/DeinitSource'
-import { DeinitMode } from '../Application/DeinitMode'
-import { AppGroupManagedApplication } from '../Application/ApplicationInterface'
+import { AppGroupManagedApplication, DeinitSource, DeinitMode } from '@standardnotes/services'
 
 export type DeinitCallback = (application: AppGroupManagedApplication, mode: DeinitMode, source: DeinitSource) => void
