+++ packages/snjs/lib/ApplicationGroup/index.ts
@@ -2,5 +2,4 @@ export * from './AppGroupCallback'
 export * from './ApplicationDescriptor'
 export * from './ApplicationGroup'
 export * from './ApplicationGroupEvent'
-export * from './DeinitCallback'
 export * from './DescriptorRecord'
