+++ packages/snjs/lib/Services/User/index.ts
@@ -1,3 +1,2 @@
-export * from './UserClientInterface'
 export * from './UserServerInterface'
 export * from './UserService'
