+++ packages/snjs/lib/Types/index.ts
@@ -1,4 +1,3 @@
 export * from './ApplicationEventPayload'
-export * from './DeinitSource'
 export * from './IconType'
 export * from './UuidString'
